# Optimize pointer alignment in utf8 validation